refactor : review-item 상단, 하단 기준으로 컴포넌트 분리
